extract to trait
